move U64 array to its own module
